requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2021.10.8=h5b45459_0
certifi=2021.5.30=py36ha15d459_0
colorama=0.4.4=pyh9f0ad1d_0
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.24=py36he2d232f_0
decorator=4.4.2=pypi_0
freetype=2.10.4=h546665d_1
icu=69.1=h0e60522_0
imageio=2.15.0=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2022.0.0=h57928b3_3663
jbig=2.1=h8d14728_2003
jpeg=9e=h8ffe710_0
kiwisolver=1.3.1=py36he95197e_1
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libblas=3.9.0=13_win64_mkl
libcblas=3.9.0=13_win64_mkl
libclang=13.0.1=default_h81446c8_0
libdeflate=1.10=h8ffe710_0
liblapack=3.9.0=13_win64_mkl
libpng=1.6.37=h1d00b33_2
libtiff=0.4.1=pypi_0
libzlib=1.2.11=h8ffe710_1014
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.3.4=py36ha15d459_0
matplotlib-base=3.3.4=py36h1abdf75_0
mkl=2022.0.0=h0e2418a_796
msys2-conda-epoch=20160418=1
networkx=2.5.1=pypi_0
nibabel=3.2.2=pypi_0
numpy=1.19.5=py36h4b40d73_2
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb211442_1
openssl=1.1.1n=h8ffe710_0
packaging=21.3=pypi_0
pillow=8.4.0=pypi_0
pip=21.3.1=pyhd8ed1ab_0
pyglet=1.5.23=pypi_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyqt=5.12.3=py36ha15d459_7
pyqt-impl=5.12.3=py36he2d232f_7
pyqt5-sip=4.19.18=py36he2d232f_7
pyqtchart=5.12=py36he2d232f_7
pyqtwebengine=5.12.1=py36he2d232f_7
python=3.6.15=h39d44d4_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.6=2_cp36m
pywavelets=1.1.1=pypi_0
qt=5.12.9=h556501e_6
rivuletpy=0.3.0=pypi_0
scikit-fmm=0.0.9=pypi_0
scikit-image=0.17.2=pypi_0
scipy=1.5.3=py36h27d303f_1
setuptools=58.0.4=py36ha15d459_2
simpleitk=2.1.1=py36ha925a31_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.37.1=h8ffe710_0
tbb=2021.5.0=h2d74725_0
tifffile=2020.9.3=pypi_0
tk=8.6.12=h8ffe710_0
tornado=6.1=py36h68aa20f_1
tqdm=4.63.1=pyhd8ed1ab_0
ucrt=10.0.20348.0=h57928b3_0
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.5=h62dcd97_1
zipp=3.6.0=pypi_0
zlib=1.2.11=h8ffe710_1014
zstd=1.5.2=h6255e5f_0

numpy~=1.19.5
scikit-fmm~=0.0.9
tqdm~=4.63.1
scipy~=1.5.3
scikit-image~=0.17.2
simpleitk~=2.0.2
six~=1.16.0
pyglet~=1.5.23
pillow~=8.4.0
rivuletpy~=0.3.0
matplotlib~=3.3.4
tifffile~=2020.9.3
setuptools~=58.0.4
vtk~=9.0.1